Updated advancedrecyclerview lib and support libs

+++ app/src/main/java/group/g203/justalittlefit/activity/ViewPastWorkout.java
@@ -91,7 +91,7 @@ public class ViewPastWorkout extends BaseNaviDrawerActivity implements ConfirmDe
                 setViewPager(workouts);
             }
         } else if (event.getResult() instanceof Integer) {
-            if (event.getResult() == workouts.size()) {
+            if ((Integer)event.getResult() == workouts.size()) {
                 Utils.displayLongToast(this, getString(R.string.workout_s_deleted));
                 Intent intent = new Intent(this, Home.class);
                 intent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
